ServidorPython/requirements.txt
django==3.2.18
django-environ==0.10.0
django-cors-headers==3.14.0
djangorestframework==3.14.0
mysqlclient==2.1.1
mysql-connector-python==8.0.32